requirements.txt
django==3.2.10
djangorestframework==3.13.1
django-allauth==0.47.0
django-rest-auth==0.9.5
django_model_utils==4.2.0
drf_yasg==1.20.0
django_environ==0.8.1
